Rename AST nodes to avoid conflicting class names
Changes adapted from trino/PR#13602
Original commit: 21695a1dfb33d5cb8bb195397d769222da5a6345
Author: kasiafi
