+++ presto-parser/src/main/java/com/facebook/presto/sql/tree/AstVisitor.java
@@ -872,12 +872,12 @@ public abstract class AstVisitor<R, C>
         return visitNode(node, context);
     }
 
-    protected R visitTableArgument(TableArgument node, C context)
+    protected R visitTableArgument(TableFunctionTableArgument node, C context)
     {
         return visitNode(node, context);
     }
 
-    protected R visitDescriptorArgument(DescriptorArgument node, C context)
+    protected R visitDescriptorArgument(TableFunctionDescriptorArgument node, C context)
     {
         return visitNode(node, context);
     }
