+++ presto-parser/src/main/java/com/facebook/presto/sql/tree/TableFunctionArgument.java
@@ -33,7 +33,7 @@ public class TableFunctionArgument
         super(Optional.of(location));
         this.name = requireNonNull(name, "name is null");
         requireNonNull(value, "value is null");
-        checkArgument(value instanceof TableArgument || value instanceof DescriptorArgument || value instanceof Expression);
+        checkArgument(value instanceof TableFunctionTableArgument || value instanceof TableFunctionDescriptorArgument || value instanceof Expression);
         this.value = value;
     }
