+++ presto-parser/src/main/java/com/facebook/presto/sql/tree/TableFunctionTableArgument.java
@@ -23,7 +23,7 @@ import java.util.stream.Collectors;
 import static com.facebook.presto.sql.ExpressionFormatter.formatSortItems;
 import static java.util.Objects.requireNonNull;
 
-public class TableArgument
+public class TableFunctionTableArgument
         extends Node
 {
     private final Relation table;
@@ -31,7 +31,7 @@ public class TableArgument
     private final Optional<OrderBy> orderBy;
     private final boolean pruneWhenEmpty;
 
-    public TableArgument(
+    public TableFunctionTableArgument(
             NodeLocation location,
             Relation table,
             Optional<List<Expression>> partitionBy,
@@ -92,7 +92,7 @@ public class TableArgument
             return false;
         }
 
-        TableArgument other = (TableArgument) o;
+        TableFunctionTableArgument other = (TableFunctionTableArgument) o;
         return Objects.equals(table, other.table) &&
                 Objects.equals(partitionBy, other.partitionBy) &&
                 Objects.equals(orderBy, other.orderBy) &&
